refactor(Erbausschlagung Anfrage): rename schema file
